+++ yash-env/CHANGELOG.md
@@ -75,6 +75,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   parameter instead of an `AtFlags` parameter.
 - The `system::System::open` method has been redefined to take `OfdAccess` and
   `OpenFlag` parameters instead of `nix::fcntl::OFlag`.
+- The `system::System::isatty` method now returns a `bool` instead of a
+  `system::Result<bool>`.
 - The `system::System::umask` method now takes and returns a value of the new
   `system::Mode` type.
 - The `system::System::sigmask` method now takes a `SigmaskOp` parameter instead